+++ site/server/src/lib/scripts/FourWinning/Rows.ts
@@ -1,0 +1,63 @@
+interface Row {
+  side: string;
+  data: string[];
+}
+
+export let rows: Row[] = [
+  {
+    side: "1 meter",
+    data: ["10-11", "12-13", "14-15", "16-17", "18-19", "20-21", "22-23"],
+  },
+  {
+    side: "1 meter",
+    data: ["24-25", "26-27", "28-29", "30-31", "32-33", "34-35", "36-37"],
+  },
+  {
+    side: "2 meters",
+    data: ["38-40", "41-43", "44-46", "47-49", "50-52", "53-55", "56-58"],
+  },
+  {
+    side: "2 meters",
+    data: ["59-61", "62-64", "65-67", "68-70", "71-73", "74-76", "77-79"],
+  },
+  {
+    side: "3 meters",
+    data: ["80-82", "83-85", "86-88", "89-91", "92-94", "95-97", "98-100"],
+  },
+  {
+    side: "3 meters",
+    data: [
+      "101-103",
+      "104-106",
+      "107-109",
+      "110-112",
+      "113-115",
+      "116-118",
+      "119-121",
+    ],
+  },
+  {
+    side: "4 meters",
+    data: [
+      "122-124",
+      "125-127",
+      "128-130",
+      "131-133",
+      "134-136",
+      "137-139",
+      "140-142",
+    ],
+  },
+  {
+    side: "4 meters",
+    data: [
+      "143-145",
+      "146-148",
+      "149-151",
+      "152-154",
+      "155-157",
+      "158-160",
+      "161-163",
+    ],
+  },
+];
